Fix #9 - ekphrastisize some poems
For ekphrastic articles, add `ekphrastic` node to YAML metadata.
This node includes subnodes `image`, `title`, `alt`, `link`, and `class`.
`image` provides a link to the local image--just include the file name
with the extension, not the folder (all images should be in /img/.)
`title` provides the title of the image, and the alt-text, if there
is no `alt` node.
`alt`, if it exists, provides the alt text for the image.
`link`, if present, wraps the image in an `<a>` tag--it should point
to the source web page of the ekphrastic image.
`class`, if present, sets the class(es) for the image, for styling.

In this commit, I've set `ekphrastic` on the four articles that have
them so far: 'The Death Zone,' 'AMBER alert,' 'The moon is gone,' and
'Man.' I've also updated .template.html with the changes, and updated
README.md to reflect the changes in YAML structure.