Mostly fix #11: Dedication/epigraph alignment
So the issue is solved in terms of how it looks, though
it adds a gross extra div into every page and uses :only-child,
which I don't think is super-supported. But it's the best I can
do that I know of until we get to better flexbox support.

Or you know, maybe later I can try doing some templating fixes--
injecting classes so that normally, .dedication is right-aligned
but when an epigraph is present, change the class to .dedication-left
or something. IDK. Either way is sort of ugly. :(