Fix #5: Verse typesetting
Thanks to the pandoc-discussion thread at
<https://groups.google.com/forum/#!topic/pandoc-discuss/_JnTJnsSK3k>,
line breaks in verse have been converted to <span class="line">s,
which enables the CSS to style them with hanging indents, given
a too-small viewport.  This commit also includes a makefile edit to
reflect this change, and the Haskell source and executable of the
pandoc filter.